Revert fetching part
