+++ src/PluginsManager/Microsoft.Performance.Toolkit.Plugins.Runtime/AvailablePlugin.cs
@@ -18,36 +18,28 @@ namespace Microsoft.Performance.Toolkit.Plugins.Runtime
     public sealed class AvailablePlugin
     {
         private readonly IPluginDiscoverer pluginDiscoverer;
-        // private readonly IPluginFetcher pluginFetcher;
+        private readonly IPluginFetcher pluginFetcher;
 
-        ///// <summary>
-        /////     Creates an instance of <see cref="AvailablePlugin"/>
-        ///// </summary>
-        ///// <param name="pluginInfo">
-        /////     The <see cref="AvailablePluginInfo"/> object cotaining information about this plugin.
-        ///// </param>
-        ///// <param name="discoverer">
-        /////     The <see cref="IPluginDiscoverer"/> this plugin is discovered by.
-        ///// </param>
-        ///// <param name="fetcher">
-        /////     The <see cref="IPluginFetcher"/> this plugin uses to fetch plugin content.
-        ///// </param>
-        //public AvailablePlugin(
-        //    AvailablePluginInfo pluginInfo,
-        //    IPluginDiscoverer discoverer,
-        //    IPluginFetcher fetcher)
-        //{
-        //    this.AvailablePluginInfo = pluginInfo;
-        //    this.pluginDiscoverer = discoverer;
-        //    this.pluginFetcher = fetcher;
-        //}
-
-        public AvailablePlugin(
+        /// <summary>
+        ///     Creates an instance of <see cref="AvailablePlugin"/>
+        /// </summary>
+        /// <param name="pluginInfo">
+        ///     The <see cref="AvailablePluginInfo"/> object cotaining information about this plugin.
+        /// </param>
+        /// <param name="discoverer">
+        ///     The <see cref="IPluginDiscoverer"/> this plugin is discovered by.
+        /// </param>
+        /// <param name="fetcher">
+        ///     The <see cref="IPluginFetcher"/> this plugin uses to fetch plugin content.
+        /// </param>
+        internal AvailablePlugin(
             AvailablePluginInfo pluginInfo,
-            IPluginDiscoverer discoverer)
+            IPluginDiscoverer discoverer,
+            IPluginFetcher fetcher)
         {
             this.AvailablePluginInfo = pluginInfo;
             this.pluginDiscoverer = discoverer;
+            this.pluginFetcher = fetcher;
         }
 
         /// <summary>
@@ -69,12 +61,12 @@ namespace Microsoft.Performance.Toolkit.Plugins.Runtime
             return this.pluginDiscoverer.GetPluginMetadataAsync(this.AvailablePluginInfo.Identity, cancellationToken);
         }
 
-        //internal Task<Stream> GetPluginPackageStream(CancellationToken cancellationToken, IProgress<int> progress)
-        //{
-        //    return this.pluginFetcher.GetPluginStreamAsync(
-        //        this.AvailablePluginInfo,
-        //        cancellationToken,
-        //        progress);
-        //}
+        public Task<Stream> GetPluginPackageStream(CancellationToken cancellationToken, IProgress<int> progress)
+        {
+            return this.pluginFetcher.GetPluginStreamAsync(
+                this.AvailablePluginInfo,
+                cancellationToken,
+                progress);
+        }
     }
 }
